DataFrameTools

Documentation for DataFrameTools.

DataFrameTools.checkparse

DataFrameTools.chknnm

DataFrameTools.convertdf2!

DataFrameTools.describeinstr

DataFrameTools.dfvstack

DataFrameTools.dropmissingcol

DataFrameTools.filtern!

DataFrameTools.filtern!

DataFrameTools.get1var

DataFrameTools.get1var

DataFrameTools.get1var

DataFrameTools.ifstrparse

DataFrameTools.isableparse

DataFrameTools.islnan

DataFrameTools.islnan

DataFrameTools.isnnm

DataFrameTools.middle

DataFrameTools.preview

DataFrameTools.readtable

DataFrameTools.reducetype

DataFrameTools.selectname

DataFrameTools.selectnames

DataFrameTools.checkparse — Method

checkparse(df::DataFrame, TYPE) check if all elements column variables can be parsed to a specific data TYPE. It returns a dictionary with the column names of df as its keys, and a vector of DataType and anything that cannot be parsed to TYPE.

DataFrameTools.chknnm — Method

chknnm(df) check if DataFrame df contains missing values or NaN. Use this before input df into machine.

DataFrameTools.convertdf2! — Method

convertdf2!(df, TYPE) convert all elements in df to the type of TYPE, ignoring missing, NaN and nothing.

DataFrameTools.describeinstr — Method

Given either missing, NaN and Number, returns "Missing", "Not a Number" and "Number".

DataFrameTools.dfvstack — Method

Imagine we have several dataframes, each have columns "t" and "f(t)", and we try to get a long dataframe of columns "t", "f(t)" and "sheet" which is specified by sheetnames; in this case, dfvstack(name_df_pairs, t_name, ft_name) can help!

Example

julia> name_df_pairs = sheetnames .=> dfs

Dict{Any, Any} with 3 entries:
  "s763.mat" => 4001×4 DataFrame…
  "s761.mat" => 1499×4 DataFrame…
  "s915.mat" => 824×4 DataFrame…

julia> t_name = "t"; 
julia> ft_name = "f(t)"
julia> longdf = dfvstack(name_df_pairs, t_name, ft_name)

DataFrameTools.dropmissingcol — Method

dropmissingcol(df) drops columns which are all missing.

DataFrameTools.filtern! — Method

filtern(col1, col2, df) automatically drop rows where either of col1 and col2 contains NaN, missing or nothing by default.

DataFrameTools.filtern! — Method

filtern works based on filter that drops rows where the corresponding specific column col1 contains NaN, missing or nothing by default.

filtern is similar to dropmissing but you can drop anything by specifying functions using kwarg chkfun, e.g. filtern(...; chkfun=(ismissing, x -> x>1)).

filtern(col1, df) is the shorthand of filter(cols => fun, df::AbstractDataFrame) where fun is for examing whether the column contains NaN, missing or nothing.

DataFrameTools.get1var — Method

get1var(df::DataFrame, expr::Regex) indexes df on the only one column that matches expr. If more than one columns are matched, it raises an error.

DataFrameTools.get1var — Method

get1var(df::DataFrame) returns the only one column of df as a Vector. If there are more than one column, it raises an error.

DataFrameTools.get1var — Method

get1var(v::Vector) get the only variable in the vector v which should have exactly one element; an error will be raised if length(v) != 1.

DataFrameTools.ifstrparse — Method

ifstrparse(x, TYPE): if x is AbstractString, then parse it to the variable of TYPE, ignoring missing, NaN and nothing (ignoring those that isnnm(x) returns true).

DataFrameTools.isableparse — Method

isableparse(x, TYPE) use try...catch to check if x can be parsed to of TYPE. If not, return x; else, return TYPE. If x::TYPE it also returns TYPE evenif it can't be parsed to the type of itself.

DataFrameTools.islnan — Method

Return true if it is literally not a number. For example, all(islnan.(["#VALUE!", "nan", "NaN", "Nan", nothing])) is true.

DataFrameTools.islnan — Method

For x being the type other than the types listed above, islnan(x) falls back to isnnm(x).

See isnnm.

DataFrameTools.isnnm — Method

Check if x is missing, nothing or NaN. Different from isnan, for x being either of Nothing, Missing, DateTime, and AbstractString, islnan(x) returns true for Nothing and Missing, and returns false for the rest.

The difference between islnan and isnnm is that, isnnm check only NaN for Not-a-Number. If you input something like "#VALUE!", "NaN", it returns false (NOT missing, nothing or NaN).

DataFrameTools.middle — Method

middle(df::DataFrame, n::Integer) Similar to first(df, n) or last(df, n), but get the dataframe of middle n rows.

DataFrameTools.preview — Method

preview(df::DataFrame, n::Integer; show_row_number=true) gives preview of first, middle and last columns of total n rows.

DataFrameTools.readtable — Method

readtable(filepath) read matlab mat or csv files, and returns a DataFrame

DataFrameTools.reducetype — Method

reducetype(v::AbstractArray) returns the array of type Union{uniqueTYPE}, where uniqueTYPE = v .|> typeof |> unique.

For more about manipulating datatype, also see nonmissingtype in missing.jl and Base.typesplit in promotion.jl

DataFrameTools.selectname — Method

selectname(df::DataFrame, arg) returns a vector of column names that match arg of the dataframe df. arg can be Regex, AbstractString, or InvertedIndex.

DataFrameTools.selectnames — Method

selectnames(dfnames, args...) select an array of names that contains arg::String or pattern that matches arg::Regex.

Example

julia> selectnames(["Soil_water_content_10cm, water_level_#1, water_level_#2"], r"water_lev")

[water_level_#1, water_level_#2"]