DataFrameTools.jl Documentation

DataFrameTools.jl Documentation

Functions

Index

Functions

DataFrameTools.file_format — Function

Returns the file format from the file suffix.

DataFrameTools.compression_formats — Function

Returns the file formats supported for compressed archives.

DataFrameTools.compress — Function

Creates a compressed file. The type of compression is inferred from the output file name. subfile is the name of the file within the compressed archive. f is a function which writes to the archive.

The typical use case is:

compress(file, subfile) do io
    # put here the body of function f(io) to write to io
end

DataFrameTools.uncompress — Function

Uncompresses a compressed file. The type of compression is inferred from the output file name. f is a function which reads from the archive.

The typical use case is:

uncompress(file) do io
    # put here the body of function f(io) to write to io
end

DataFrameTools.df_formats — Function

Returns the file formats supported for DataFrames.

DataFrameTools.df_write — Function

Writes a DataFrame to a file. The file suffix determines how the DataFrame is serialized. If the file has a compressed suffix, subformat determines how the DataFrame is serialized in the compressed archive.

subformat::Union{Nothing,Symbol}=nothing: format to used within a compressed archive.

dictencode::Bool=true: true to use a dictionary encoding, if possible, and false otherwise.

DataFrameTools.df_read — Function

Reads a DataFrame from a file. The file suffix determines how the DataFrame is deserialized.

Index

DataFrameTools.compress

DataFrameTools.compression_formats

DataFrameTools.df_formats

DataFrameTools.df_read

DataFrameTools.df_write

DataFrameTools.file_format

DataFrameTools.uncompress